[code]
Fix bug 5719. The ui part is now able to handle double quotes (un)escapes transparently and automatically.

The user now interacts with the unescaped string version. When the edited string returns back to the doxyfile, it is automatically escaped.

A preference allows the user to deactivate this bahviour (default is on).